+++ src/test/scala/matrix.scala
@@ -33,7 +33,7 @@ class MatrixTest extends FunSuite with BeforeAndAfter
    ***************************************************************************/
   test("Calculate 1x1 matrix and vector") {
     val cm = new Matrix(sc.parallelize(List( (1,(1,1)) )), sc.parallelize(List()))
-    val ar = sc.parallelize(List[(Int,Double)]( (1,-2) ))
+    val ar = sc.parallelize(List[(Long,Double)]( (1,-2) ))
     assert( (cm*ar).collect.sorted === ar.collect.sorted )
   }
   // careful that the matrix format is (from,(to,entry))
@@ -44,7 +44,7 @@ class MatrixTest extends FunSuite with BeforeAndAfter
       (1,(1,1)),
       (2,(2,1))
     )), sc.parallelize(List()))
-    val ar = sc.parallelize(List[(Int,Double)]( (1,-2), (2,7) ))
+    val ar = sc.parallelize(List[(Long,Double)]( (1,-2), (2,7) ))
     assert( (cm*ar).collect.sorted === ar.collect.sorted )
   }
 
@@ -53,8 +53,8 @@ class MatrixTest extends FunSuite with BeforeAndAfter
       (2,(1,1)),
       (1,(2,1))
     )), sc.parallelize(List()))
-    val ar = sc.parallelize(List[(Int,Double)]( (1,-2), (2,7) ))
-    val pr = sc.parallelize(List[(Int,Double)]( (1,7), (2,-2) ))
+    val ar = sc.parallelize(List[(Long,Double)]( (1,-2), (2,7) ))
+    val pr = sc.parallelize(List[(Long,Double)]( (1,7), (2,-2) ))
     assert( (cm*ar).collect.sorted === pr.collect.sorted )
   }
 
@@ -63,8 +63,8 @@ class MatrixTest extends FunSuite with BeforeAndAfter
       (1,(1,1)), (2,(1,2)),
       (1,(2,3)), (2,(2,4))
     )), sc.parallelize(List()))
-    val ar = sc.parallelize(List[(Int,Double)]( (1,5), (2,6) ))
-    val pr = sc.parallelize(List[(Int,Double)]( (1,17), (2,39) ))
+    val ar = sc.parallelize(List[(Long,Double)]( (1,5), (2,6) ))
+    val pr = sc.parallelize(List[(Long,Double)]( (1,17), (2,39) ))
     assert( (cm*ar).collect.sorted === pr.collect.sorted )
   }
 
@@ -74,8 +74,8 @@ class MatrixTest extends FunSuite with BeforeAndAfter
       (1,(2,2)),
       (2,(2,3))
     )), sc.parallelize(List()))
-    val ar = sc.parallelize(List[(Int,Double)]( (1,1), (2,2), (3,3) ))
-    val pr = sc.parallelize(List[(Int,Double)]( (1,2), (2,8), (3,0) ))
+    val ar = sc.parallelize(List[(Long,Double)]( (1,1), (2,2), (3,3) ))
+    val pr = sc.parallelize(List[(Long,Double)]( (1,2), (2,8), (3,0) ))
     assert( (cm*ar).collect.sorted === pr.collect.sorted )
   }
 
@@ -85,8 +85,8 @@ class MatrixTest extends FunSuite with BeforeAndAfter
       (1,(2,1)),
       (3,(3,1))
     )), sc.parallelize(List()))
-    val ar = sc.parallelize(List[(Int,Double)]( (1,1), (2,2), (3,3) ))
-    val pr = sc.parallelize(List[(Int,Double)]( (1,2), (2,1), (3,3) ))
+    val ar = sc.parallelize(List[(Long,Double)]( (1,1), (2,2), (3,3) ))
+    val pr = sc.parallelize(List[(Long,Double)]( (1,2), (2,1), (3,3) ))
     assert( (cm*ar).collect.sorted === pr.collect.sorted )
   }
 
@@ -98,8 +98,8 @@ class MatrixTest extends FunSuite with BeforeAndAfter
       (3,(2,-2)),
       (2,(3,3))
     )), sc.parallelize(List()))
-    val ar = sc.parallelize(List[(Int,Double)]( (1,7), (2,0), (3,-2) ))
-    val pr = sc.parallelize(List[(Int,Double)]( (1,-63), (2,11), (3,0) ))
+    val ar = sc.parallelize(List[(Long,Double)]( (1,7), (2,0), (3,-2) ))
+    val pr = sc.parallelize(List[(Long,Double)]( (1,-63), (2,11), (3,0) ))
     assert( (cm*ar).collect.sorted === pr.collect.sorted )
   }
 
@@ -111,24 +111,24 @@ class MatrixTest extends FunSuite with BeforeAndAfter
       (3,(2,4)),
       (4,(2,2))
     )), sc.parallelize(List()))
-    val ar = sc.parallelize(List[(Int,Double)]( (1,-4), (2,2), (3,8), (4,0) ))
-    val pr = sc.parallelize(List[(Int,Double)]( (1,20), (2,28), (3,0), (4,0) ))
+    val ar = sc.parallelize(List[(Long,Double)]( (1,-4), (2,2), (3,8), (4,0) ))
+    val pr = sc.parallelize(List[(Long,Double)]( (1,20), (2,28), (3,0), (4,0) ))
     assert( (cm*ar).collect.sorted === pr.collect.sorted )
   }
 
   test("Calculate 1x1 matrix with constant column optimization") {
     val cm = new Matrix(sc.parallelize(List()),
-      sc.parallelize(List[(Int,Double)]((1,1.0))))
-    val ar = sc.parallelize(List[(Int,Double)]( (1,0.7) ))
+      sc.parallelize(List[(Long,Double)]((1,1.0))))
+    val ar = sc.parallelize(List[(Long,Double)]( (1,0.7) ))
     assert( (cm*ar).collect.sorted === ar.collect.sorted )
   }
 
   test("Calculate 2x2 matrix with constant column optimization") {
     val cm = new Matrix(sc.parallelize(List(
       (1,(2,1))
-    )), sc.parallelize(List[(Int,Double)]((2,0.5))))
-    val ar = sc.parallelize(List[(Int,Double)]( (1,0.5), (2,0.5) ))
-    val pr = sc.parallelize(List[(Int,Double)]( (1,0.25), (2,0.75) ))
+    )), sc.parallelize(List[(Long,Double)]((2,0.5))))
+    val ar = sc.parallelize(List[(Long,Double)]( (1,0.5), (2,0.5) ))
+    val pr = sc.parallelize(List[(Long,Double)]( (1,0.25), (2,0.75) ))
     assert( (cm*ar).collect.sorted === pr.collect.sorted )
   }
